gcp_hosting/cloud_function/requirements.txt
functions-framework==3.*
great_expectations==0.15.41
sqlalchemy-bigquery==1.4.4
google-cloud-storage==2.5.0